The turrets comment is back on turrets, and clippy has thirty fewer things to say
`cargo clippy --fix` took the mechanical thirty: the borrows the compiler
immediately dereferences, the `% n == 0`s that are `is_multiple_of`, the two
`vec!`s that were arrays, the casts to their own type, the `?` written out
long-hand, and the three test modules that had items after them — moved to the
end of their files, which is where a test module belongs. The seven
`needless_range_loop`s were hand-done, because the fix is not an iterator but
`fill` over a slice, and the three `Draft::default()` tests now say what they
change in the initialiser.

**The one real defect here is a comment on the wrong function.** The paragraph
that says a turret list is sorted because a `HashMap` iterates differently in
different processes — the comment simplifying.md holds up as the model of one
worth keeping — had come unglued from `turrets` and was sitting on the front of
`blast_centre`'s doc, so the function it warns about had no warning on it and
the function it did not describe had two paragraphs about turrets. It is back
where it belongs.

Both `#[allow]`s in the tree were stale and are gone: `shadowed_in` has seven
parameters and the lint fires at eight, and the binding under
`let_and_return` is mutated before it is returned, so neither could ever have
fired. Clippy confirms nothing new appeared without them.

Forty-two warnings to twelve, all six gates green, 577 tests and 388 without
the renderer, unchanged either way.

+++ src/server/profiles.rs
@@ -390,7 +390,7 @@ mod tests {
     fn a_search_answers_no_more_than_it_is_asked_for() {
         let mut store = Profiles::new();
         for i in 0..40 {
-            rated(&mut store, &format!("p{i:02}"), 1000 + i as i32);
+            rated(&mut store, &format!("p{i:02}"), 1000 + i);
         }
         assert_eq!(store.search("", crate::net::PEOPLE_MOST).len(), crate::net::PEOPLE_MOST);
         assert_eq!(store.search("p", crate::net::PEOPLE_MOST).len(), crate::net::PEOPLE_MOST);
